refactor: move MSE grid search functions to mse_quant.py
